Move runner to its own module
